Revert "Refactor component locations"
