+++ src/App.js
@@ -5,8 +5,7 @@ import { onAuthStateChanged } from "firebase/auth";
 import { setUser, removeUser } from "./store/slices/userSlice";
 import { checkPremiumStatus } from "./components/Premium/services/premiumUtils";
 
-import Header from "./components/Header/Header";
-import Footer from "./components/Footer/Footer";
+import Header from "./pages/Header/Header";
 import AppRoutes from "./routes/AppRoutes";
 import Loader from "./components/Loader/Loader";
 import "./App.css";
@@ -42,12 +41,11 @@ const App = () => {
 
   return (
     <div className="container-fluid p-0">
-      <div className="w-100 d-flex flex-column min-vh-100">
+      <div className="w-100">
         <Header />
-        <div className="container-custom flex-grow-1">
+        <div className="container-custom">
           <AppRoutes />
         </div>
-        <Footer />
       </div>
     </div>
   );
